Split the Pure Data instance into its own node
Should make having multiple audio stream players easier
