+++ src/register_types.cpp
@@ -1,5 +1,6 @@
 #include "register_types.h"
 
+#include "pure_data_instance.h"
 #include "pure_data_audio_stream_player.h"
 #include "pure_data_patch.h"
 
@@ -14,6 +15,7 @@ void initialize_example_module(ModuleInitializationLevel p_level) {
 		return;
 	}
 
+	ClassDB::register_class<PureDataInstance>();
 	ClassDB::register_class<PureDataAudioStream>();
 	ClassDB::register_class<PureDataPatch>();
 }
